requirements.txt
streamlit
langchain
cohere
pinecone-client
bs4
requests
PyPDF2
python-docx
